+++ src/main/java/com/example/planservice/repository/GoalRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface GoalRepository extends JpaRepository<Goal,Long> {
+public interface GoalRepository extends JpaRepository<Goal,Integer> {
 
     List<Goal> findByName(String name);
 }
